% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyThresholds.R
\name{applyThresholds}
\alias{applyThresholds}
\title{Use thresholds to decide whether a taxon is classed as positive or negative}
\usage{
applyThresholds(
  type,
  rpm_ratio,
  reads,
  control_reads,
  proportion,
  thresholds,
  reads_thres_bac = 0,
  reads_thres_vir = 0,
  reads_thres_fun = 0,
  reads_thres_euk = 0,
  rpm_ratio_thres_bac = 0,
  rpm_ratio_thres_vir = 0,
  rpm_ratio_thres_fun = 0,
  rpm_ratio_thres_euk = 0,
  proportion_thres_bac = 0,
  proportion_thres_vir = 0,
  proportion_thres_fun = 0,
  proportion_thres_euk = 0,
  low_level_bac = FALSE,
  low_level_vir = FALSE,
  low_level_fun = FALSE,
  low_level_euk = FALSE
)
}
\value{
Positive or negative
}
\description{
Use thresholds to decide whether a taxon is classed as positive or negative
}
