% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyThresholdsReport.R
\name{applyThresholdsReport}
\alias{applyThresholdsReport}
\title{Link sample data to corresponding controls}
\usage{
applyThresholdsReport(
  report,
  thresholds_filepath = NULL,
  proportion = "proportion_nonhuman_classified",
  reads_thres_bac = 0,
  reads_thres_vir = 0,
  reads_thres_fun = 0,
  reads_thres_euk = 0,
  rpm_ratio_thres_bac = 0,
  rpm_ratio_thres_vir = 0,
  rpm_ratio_thres_fun = 0,
  rpm_ratio_thres_euk = 0,
  proportion_thres_bac = 0,
  proportion_thres_vir = 0,
  proportion_thres_fun = 0,
  proportion_thres_euk = 0,
  low_level_bac = FALSE,
  low_level_vir = FALSE,
  low_level_fun = FALSE,
  low_level_euk = FALSE
)
}
\arguments{
\item{report_full}{reports linked with relevant control IDs}
}
\value{
A report with control read counts linked to their relevant sample
}
\description{
Link sample data to corresponding controls
}
