% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseSpecies.R
\name{collapseSpecies}
\alias{collapseSpecies}
\title{Wrapper function to calculate reads per million ratio and proportion of micorbial reads from raw classifier output}
\usage{
collapseSpecies(report_full, db)
}
\arguments{
\item{report_full}{reports combined with db and sample info}

\item{db}{output of makeTaxDBReports}
}
\value{
Report collapsed by species
}
\description{
Wrapper function to calculate reads per million ratio and proportion of micorbial reads from raw classifier output
}
