% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareControls.R
\name{compareControls}
\alias{compareControls}
\title{Create species-level database}
\usage{
compareControls(report_full)
}
\arguments{
\item{taxid}{Taxon ID}

\item{taxonomizr_sql}{sql fro taxonomizr}
}
\value{
A named list containing fileds associated with the input taxon ID
}
\description{
Create species-level database
}
