% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSpecies.R
\name{countSpecies}
\alias{countSpecies}
\title{Count true and false positive species present in samples and calculate sensitivity and precision}
\usage{
countSpecies(
  combined_dnarna_report,
  positive_species_filepath,
  by_type = FALSE
)
}
\arguments{
\item{combined_dnarna_report}{output of combineDNARNA}

\item{positive_species_filepath}{Path to csv file with list of expected positive taxon IDs}

\item{by_type}{if true, calculates stas per type (e.g. bacteria, virus) within each sample. Default false.}
}
\value{
A report with a row for each sample with stats calculated
}
\description{
Count true and false positive species present in samples and calculate sensitivity and precision
}
