% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractThreshold.R
\name{extractThreshold}
\alias{extractThreshold}
\title{Extract thresholds from thresholds csv}
\usage{
extractThreshold(thresholds, threshold_name)
}
\arguments{
\item{thresholds}{imported thresholds.csv}

\item{threshold_name}{descriptive name of threshold}
}
\value{
A report with control read counts linke dto their relevant sample
}
\description{
Extract thresholds from thresholds csv
}
