% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAllReports.R
\name{importAllReports}
\alias{importAllReports}
\title{Import all reports}
\usage{
importAllReports(samplesheet)
}
\arguments{
\item{samplesheet}{samplesheet}
}
\value{
A report for all the samples in the samplesheet
}
\description{
Import all reports
}
