% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importReport.R
\name{importReport}
\alias{importReport}
\title{import outputs of taxonomic classifiers}
\usage{
importReport(sample_id, tool, input_filepath)
}
\arguments{
\item{sample_id}{name of your sample}

\item{tool}{program used to generate report}

\item{input_filepath}{filepath to report corresponding to sample_id and tool}
}
\value{
A report containing just the taxids and read counts
}
\description{
import outputs of taxonomic classifiers
}
