% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSamplesheet.R
\name{importSamplesheet}
\alias{importSamplesheet}
\title{Import samplesheet}
\usage{
importSamplesheet(samplesheet_filepath)
}
\arguments{
\item{samplesheet_filepath}{filepath to samplesheet}
}
\value{
A named list containing fieds associated with the input taxon ID
}
\description{
Import samplesheet
}
