% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFullReport.R
\name{makeFullReport}
\alias{makeFullReport}
\title{Wrapper function to use provided thresholds to filter the raw classifier outputs provided in the samplesheet.}
\usage{
makeFullReport(
  samplesheet_filepath = "./samplesheet.csv",
  taxonomizr_sql = "./nameNode.sqlite",
  db_filepath = NA,
  thresholds_filepath = "./thresholds.csv",
  proportion = "proportion_nonhuman_classified",
  positive_species_filepath = NULL,
  reads_thres_bac = 0,
  reads_thres_vir = 0,
  reads_thres_fun = 0,
  reads_thres_euk = 0,
  rpm_ratio_thres_bac = 0,
  rpm_ratio_thres_vir = 0,
  rpm_ratio_thres_fun = 0,
  rpm_ratio_thres_euk = 0,
  proportion_thres_bac = 0,
  proportion_thres_vir = 0,
  proportion_thres_fun = 0,
  proportion_thres_euk = 0,
  low_level_bac = FALSE,
  low_level_vir = FALSE,
  low_level_fun = FALSE,
  low_level_euk = FALSE,
  collapse_species = TRUE,
  keep_species_only = FALSE,
  virus_only = FALSE
)
}
\arguments{
\item{samplesheet_filepath}{Path to samplesheet. See https://github.com/sarah-buddle/metathresholds for a template.}

\item{taxonomizr_sql}{Path to nameNode.sqlite file for taxonomizr. If this doesn't exist, a new one will be created in this location. Generating this file may take some time.}

\item{db_filepath}{Path to taxonomy database used by metaThresholds. If this doesn't exist, a new one will be created in this location. If no filepath is provided, a new database will be generated for each analysis.}

\item{thresholds_filepath}{Path to file containing thresholds. See https://github.com/sarah-buddle/metathresholds for a template.}

\item{proportion}{Type of proportion used for calculation of proportion thresholds. Options: "proportion_nonhuman_classified" (default) or "proportion_raw"}

\item{positive_species_filepath}{Path to csv file with list of expected positive taxon IDs}

\item{reads_thres_bac}{see readme}

\item{reads_thres_vir}{see readme}

\item{reads_thres_fun}{see readme}

\item{reads_thres_euk}{see readme}

\item{rpm_ratio_thres_bac}{see readme}

\item{rpm_ratio_thres_vir}{see readme}

\item{rpm_ratio_thres_fun}{see readme}

\item{rpm_ratio_thres_euk}{see readme}

\item{proportion_thres_vir}{see readme}

\item{proportion_thres_fun}{see readme}

\item{proportion_thres_euk}{see readme}

\item{low_level_bac}{see readme}

\item{low_level_vir}{see readme}

\item{low_level_fun}{see readme}

\item{low_level_euk}{see readme}

\item{collapse_species}{if true, any reads assigned to sub-species, strains etc will be assigned to the relevant species for analysis. Defaut true}

\item{keep_species_only}{if true, ooly keep taxa with the rank species in the results, Default false}

\item{virus_only}{if true, only keep viruses in the final output. Default false}

\item{proportion_thres_bas}{see readme}
}
\value{
Report with taxons classified as positive and negative according to thresholds.
}
\description{
Wrapper function to use provided thresholds to filter the raw classifier outputs provided in the samplesheet.
}
