% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTaxDB.R
\name{makeTaxDB}
\alias{makeTaxDB}
\title{Create species-level database}
\usage{
makeTaxDB(taxid, taxonomizr_sql)
}
\arguments{
\item{taxid}{Taxon ID}

\item{taxonomizr_sql}{sql fro taxonomizr}
}
\value{
A named list containing fileds associated with the input taxon ID
}
\description{
Create species-level database
}
