% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTaxDBReports.R
\name{makeTaxDBReports}
\alias{makeTaxDBReports}
\title{Create species-level database}
\usage{
makeTaxDBReports(reports, taxonomizr_sql, db_filepath = NULL)
}
\arguments{
\item{reports}{reports to create db for}

\item{taxonomizr_sql}{sqlite file for taxonomizr}

\item{db_filepath}{If provided, loads and saves existig db rather than making one from scratch}
}
\value{
A named list containing fileds associated with the input taxon ID
}
\description{
Create species-level database
}
