% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProportions.R
\name{readProportions}
\alias{readProportions}
\title{Create species-level database}
\usage{
readProportions(report_controls, contaminant_taxids)
}
\arguments{
\item{report_controls}{output of linkControls}

\item{contaminant_taxids}{list of contaminant taxids to remove}
}
\value{
A named list containing fileds associated with the input taxon ID
}
\description{
Create species-level database
}
