% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyThresholds.R
\name{applyThresholds}
\alias{applyThresholds}
\title{Use thresholds to decide whether a taxon is classed as positive or negative}
\usage{
applyThresholds(
  type,
  rpm_ratio,
  reads,
  control_reads,
  proportion,
  thresholds,
  reads_thres_bac = 0,
  reads_thres_vir = 0,
  reads_thres_fun = 0,
  reads_thres_euk = 0,
  rpm_ratio_thres_bac = 0,
  rpm_ratio_thres_vir = 0,
  rpm_ratio_thres_fun = 0,
  rpm_ratio_thres_euk = 0,
  proportion_thres_bac = 0,
  proportion_thres_vir = 0,
  proportion_thres_fun = 0,
  proportion_thres_euk = 0,
  low_level_bac = FALSE,
  low_level_vir = FALSE,
  low_level_fun = FALSE,
  low_level_euk = FALSE
)
}
\arguments{
\item{type}{options: bacteria, virus, fungi, other_eukaryote}

\item{rpm_ratio}{reads per million ratio}

\item{reads}{number ofreads}

\item{control_reads}{number of reads in negative control}

\item{proportion}{proportion}

\item{thresholds}{thresholds imported from thresholds_filepath}

\item{reads_thres_bac}{see readme}

\item{reads_thres_vir}{see readme}

\item{reads_thres_fun}{see readme}

\item{reads_thres_euk}{see readme}

\item{rpm_ratio_thres_bac}{see readme}

\item{rpm_ratio_thres_vir}{see readme}

\item{rpm_ratio_thres_fun}{see readme}

\item{rpm_ratio_thres_euk}{see readme}

\item{proportion_thres_vir}{see readme}

\item{proportion_thres_fun}{see readme}

\item{proportion_thres_euk}{see readme}

\item{low_level_bac}{see readme}

\item{low_level_vir}{see readme}

\item{low_level_fun}{see readme}

\item{low_level_euk}{see readme}

\item{proportion_thres_bas}{see readme}
}
\value{
Positive or negative
}
\description{
Use thresholds to decide whether a taxon is classed as positive or negative
}
