% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyThresholdsReport.R
\name{applyThresholdsReport}
\alias{applyThresholdsReport}
\title{Link sample data to corresponding controls}
\usage{
applyThresholdsReport(
  report,
  thresholds_filepath = NULL,
  proportion = "proportion_nonhuman_classified",
  reads_thres_bac = 0,
  reads_thres_vir = 0,
  reads_thres_fun = 0,
  reads_thres_euk = 0,
  rpm_ratio_thres_bac = 0,
  rpm_ratio_thres_vir = 0,
  rpm_ratio_thres_fun = 0,
  rpm_ratio_thres_euk = 0,
  proportion_thres_bac = 0,
  proportion_thres_vir = 0,
  proportion_thres_fun = 0,
  proportion_thres_euk = 0,
  low_level_bac = FALSE,
  low_level_vir = FALSE,
  low_level_fun = FALSE,
  low_level_euk = FALSE
)
}
\arguments{
\item{report}{reports linked with relevant control IDs}

\item{thresholds_filepath}{Path to file containing thresholds. See https://github.com/sarah-buddle/metathresholds for a template.}

\item{proportion}{Type of proportion used for calculation of proportion thresholds. Options: "proportion_nonhuman_classified" (default) or "proportion_raw"}

\item{reads_thres_bac}{see readme}

\item{reads_thres_vir}{see readme}

\item{reads_thres_fun}{see readme}

\item{reads_thres_euk}{see readme}

\item{rpm_ratio_thres_bac}{see readme}

\item{rpm_ratio_thres_vir}{see readme}

\item{rpm_ratio_thres_fun}{see readme}

\item{rpm_ratio_thres_euk}{see readme}

\item{proportion_thres_vir}{see readme}

\item{proportion_thres_fun}{see readme}

\item{proportion_thres_euk}{see readme}

\item{low_level_bac}{see readme}

\item{low_level_vir}{see readme}

\item{low_level_fun}{see readme}

\item{low_level_euk}{see readme}

\item{proportion_thres_bas}{see readme}
}
\value{
A report with control read counts linked to their relevant sample
}
\description{
Link sample data to corresponding controls
}
