% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseSpecies.R
\name{collapseSpecies}
\alias{collapseSpecies}
\title{Transfers all reads assigned to subspecies to their corresponding species}
\usage{
collapseSpecies(report_full, db)
}
\arguments{
\item{report_full}{reports combined with db and sample info}

\item{db}{output of makeTaxDBReports}
}
\value{
Report collapsed by species
}
\description{
Transfers all reads assigned to subspecies to their corresponding species
}
