% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineDNARNA.R
\name{combineDNARNA}
\alias{combineDNARNA}
\title{Combine DNA and RNA results, retaining only whichever method gives a higher rpm}
\usage{
combineDNARNA(report)
}
\arguments{
\item{report}{Taxon ID}
}
\value{
Combine DNA and RNA report
}
\description{
Combine DNA and RNA results, retaining only whichever method gives a higher rpm
}
