% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractThreshold.R
\name{extractThreshold}
\alias{extractThreshold}
\title{Link sample data to corresponding controls}
\usage{
extractThreshold(thresholds, threshold_name)
}
\arguments{
\item{report_full}{reports linked with relevant control IDs}
}
\value{
A report with control read counts linke dto their relevant sample
}
\description{
Link sample data to corresponding controls
}
