% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importReport.R
\name{importReport}
\alias{importReport}
\title{Create species-level database}
\usage{
importReport(sample_id, tool, input_filepath)
}
\arguments{
\item{sample_id}{name of your sample}

\item{tool}{program used to generate report - can be kraken2}

\item{input_filepath}{filepath to report corresponding to sample_id and tool}
}
\value{
A report containing just the taxids and read counts
}
\description{
Create species-level database
}
