% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCorrect.R
\name{isCorrect}
\alias{isCorrect}
\title{Create species-level database}
\usage{
isCorrect(taxid, rank, result, positive_species_taxids)
}
\arguments{
\item{taxid}{Taxon ID}

\item{positive_taxids}{list of true positive taxids}
}
\value{
A named list which will become the taxonomy database entry associated with the input taxon ID.
}
\description{
Create species-level database
}
