% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProportions.R
\name{readProportions}
\alias{readProportions}
\title{Calculate proportions of total, nonhuman classified and reads of that type}
\usage{
readProportions(report_controls, contaminant_taxids = NULL)
}
\arguments{
\item{report_controls}{output of linkControls}

\item{contaminant_taxids}{list of contaminant taxids to remove}
}
\value{
A named list containing fileds associated with the input taxon ID
}
\description{
Calculate proportions of total, nonhuman classified and reads of that type
}
